move structure like android type
