+++ server/src/main/kotlin/com/example/data/entity/UserEntity.kt
