+++ server/src/main/kotlin/com/example/data/response/AuthResponse.kt
