+++ server/src/main/kotlin/com/example/plugins/Security.kt
@@ -5,10 +5,7 @@ import io.ktor.server.auth.jwt.*
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
 import com.example.token.TokenConfig
-import io.ktor.server.sessions.*
-import io.ktor.server.response.*
 import io.ktor.server.application.*
-import io.ktor.server.routing.*
 
 fun Application.configureSecurity(
     tokenConfig: TokenConfig
