+++ server/src/main/kotlin/com/example/routing/AuthRoutes.kt
