+++ server/src/main/resources/application.conf
@@ -4,7 +4,7 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ com.example.ApplicationKt.module ]
+        modules = [ ApplicationKt.module ]
     }
 }
 jwt {
