+++ settings.gradle.kts
@@ -13,3 +13,4 @@ dependencyResolutionManagement {
     }
 }
 rootProject.name = "ktor-jwt-sample"
+include(":server")
